fix some remaining warnings - fix unused nullEventId warning - make Profiler::m_stackPos unsigned, to avoid warning (and because it should never be negative) - fix signed/unsigned comparison in assert (safe because in branch where checked >= 0) - TransfomationMatrix::insertRotateOp - avoid possibly using uninitialized op - Diffcore: remove unused variable in __AVX2__ branch

+++ lib/AL_USDMaya/AL/usdmaya/nodes/RendererManager.cpp
@@ -272,7 +272,7 @@ void RendererManager::changeRendererPlugin(ProxyShape* proxy, bool creation)
       if (rendererId == 0 && creation)
         return;
       
-      assert(rendererId < m_rendererPluginsTokens.size());
+      assert(static_cast<size_t>(rendererId) < m_rendererPluginsTokens.size());
       TfToken plugin = m_rendererPluginsTokens[rendererId];
       if (!proxy->engine()->SetRendererPlugin(plugin))
       {
